added
A Python task reminder programme is a piece of software created to assist users in keeping track of their appointments and tasks. The user is prompted to enter tasks or appointments together with the associated times and dates. The programme then saves this information in a data structure such a list or dictionary.

The task list can be displayed chronologically, or the user can organise the tasks by priority or category. Depending on the user's selected notification settings, the program can also send reminders to them by email or text message.

A task reminder program written in Python might potentially include more complex capabilities like the capacity to establish recurring tasks, attach files or comments to tasks, and incorporate